+++ sdk/monitor/azure-monitor-query/tests/async/test_metrics_client_async.py
@@ -2,7 +2,7 @@ from datetime import datetime, time, timedelta
 import pytest
 import os
 from azure.identity.aio import ClientSecretCredential
-from azure.monitor.query import AggregationType
+from azure.monitor.query import MetricAggregationType
 from azure.monitor.query.aio import MetricsQueryClient
 
 def _credential():
@@ -14,6 +14,7 @@ def _credential():
     return credential
 
 @pytest.mark.live_test_only
+@pytest.mark.asyncio
 async def test_metrics_auth():
     credential = _credential()
     client = MetricsQueryClient(credential)
@@ -21,10 +22,25 @@ async def test_metrics_auth():
         os.environ['METRICS_RESOURCE_URI'],
         metric_names=["MatchedEventCount"],
         timespan=timedelta(days=1),
-        aggregations=[AggregationType.COUNT]
+        aggregations=[MetricAggregationType.COUNT]
         )
     assert response
     assert response.metrics
+
+@pytest.mark.live_test_only
+@pytest.mark.asyncio
+async def test_metrics_granularity():
+    credential = _credential()
+    client = MetricsQueryClient(credential)
+    response = await client.query(
+        os.environ['METRICS_RESOURCE_URI'],
+        metric_names=["MatchedEventCount"],
+        timespan=timedelta(days=1),
+        granularity=timedelta(minutes=5),
+        aggregations=[MetricAggregationType.COUNT]
+        )
+    assert response
+    assert response.granularity == timedelta(minutes=5)
 
 @pytest.mark.live_test_only
 async def test_metrics_namespaces():