+++ sdk/monitor/azure-monitor-query/tests/test_metrics_client.py
@@ -2,7 +2,7 @@ import pytest
 import os
 from datetime import datetime, timedelta
 from azure.identity import ClientSecretCredential
-from azure.monitor.query import MetricsQueryClient, AggregationType
+from azure.monitor.query import MetricsQueryClient, MetricAggregationType
 
 def _credential():
     credential  = ClientSecretCredential(
@@ -20,10 +20,24 @@ def test_metrics_auth():
         os.environ['METRICS_RESOURCE_URI'],
         metric_names=["MatchedEventCount"],
         timespan=timedelta(days=1),
-        aggregations=[AggregationType.COUNT]
+        aggregations=[MetricAggregationType.COUNT]
         )
     assert response
     assert response.metrics
+
+@pytest.mark.live_test_only
+def test_metrics_granularity():
+    credential = _credential()
+    client = MetricsQueryClient(credential)
+    response = client.query(
+        os.environ['METRICS_RESOURCE_URI'],
+        metric_names=["MatchedEventCount"],
+        timespan=timedelta(days=1),
+        granularity=timedelta(minutes=5),
+        aggregations=[MetricAggregationType.COUNT]
+        )
+    assert response
+    assert response.granularity == timedelta(minutes=5)
 
 @pytest.mark.live_test_only
 def test_metrics_namespaces():