+++ sdk/monitor/azure-monitor-query/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ### Features Added
 
+- Added additional `display_description` attribute to the `Metric` type.
+- Added a `MetricClass` enum to provide the class of a metric.
+- Added a `metric_class` attribute to the `MetricDefinition` type.
+- Added a `MetricNamespaceClassification` enum to support the `namespace_classification` attribute on `MetricNamespace` type.
+
 ### Breaking Changes
 
 - Rename `batch_query` to `query_batch`.
@@ -15,8 +20,16 @@
 - `top` is renamed to `max_results` in the metric's `query` API.
 - `metric_namespace_name` is renamed to `fully_qualified_namespace`
 - `is_dimension_required` is renamed to `dimension_required`
-- `time_grain` is renamed to `granularity`
+- `interval`  and `time_grain` are renamed to `granularity`
+- `orderby` is renamed to `order_by`
 - `LogsQueryResult` now returns `datetime` objects for a time values.
+- `LogsBatchQuery` doesn't accept a `request_id` anymore.
+- `MetricsMetadataValues` is removed. A dictionary is used instead.
+- `time_stamp` is renamed to `timestamp` in `MetricValue` type.
+- `AggregationType` is renamed to `MetricAggregationType`.
+- Removed `LogsBatchResultError` type.
+- `LogsQueryResultTable` is named to `LogsTable`
+- `LogsQueryResultColumn` is renamed to `LogsTableColumn`
 
 ### Bugs Fixed